Remove BigQuery::View
Move all View methods except View#data to Table.

Breaking changes: Remove View class and View#data

Keep Dataset#create_view but change return type to Table.

[refs #1785]
