+++ google-cloud-bigquery/lib/google/cloud/bigquery/dataset/access.rb
@@ -185,8 +185,8 @@ module Google
           ##
           # Add reader access to a view.
           #
-          # @param [Google::Cloud::Bigquery::View, String] view A view object or
-          #   a string identifier as specified by the [Query
+          # @param [Google::Cloud::Bigquery::Table, String] view A table object
+          #   or a string identifier as specified by the [Query
           #   Reference](https://cloud.google.com/bigquery/query-reference#from):
           #   `project_name:datasetId.tableId`.
           #
@@ -444,8 +444,8 @@ module Google
           ##
           # Remove reader access from a view.
           #
-          # @param [Google::Cloud::Bigquery::View, String] view A view object or
-          #   a string identifier as specified by the [Query
+          # @param [Google::Cloud::Bigquery::Table, String] view A table object
+          #   or a string identifier as specified by the [Query
           #   Reference](https://cloud.google.com/bigquery/query-reference#from):
           #   `project_name:datasetId.tableId`.
           #
@@ -699,8 +699,8 @@ module Google
           ##
           # Checks reader access for a view.
           #
-          # @param [Google::Cloud::Bigquery::View, String] view A view object or
-          #   a string identifier as specified by the [Query
+          # @param [Google::Cloud::Bigquery::Table, String] view A table object
+          #   or a string identifier as specified by the [Query
           #   Reference](https://cloud.google.com/bigquery/query-reference#from):
           #   `project_name:datasetId.tableId`.
           #
