+++ google-cloud-bigquery/test/google/cloud/bigquery/dataset_reference_test.rb
@@ -107,7 +107,7 @@ describe Google::Cloud::Bigquery::Dataset, :reference, :mock_bigquery do
 
     mock.verify
 
-    table.must_be_kind_of Google::Cloud::Bigquery::View
+    table.must_be_kind_of Google::Cloud::Bigquery::Table
     table.table_id.must_equal view_id
     table.query.must_equal query
     table.must_be :query_standard_sql?
