+++ google-cloud-bigquery/test/google/cloud/bigquery/view_update_test.rb
@@ -14,14 +14,14 @@
 
 require "helper"
 
-describe Google::Cloud::Bigquery::View, :update, :mock_bigquery do
+describe Google::Cloud::Bigquery::Table, :view, :update, :mock_bigquery do
   let(:dataset_id) { "my_dataset" }
   let(:table_id) { "my_view" }
   let(:table_name) { "My View" }
   let(:description) { "This is my view" }
   let(:view_hash) { random_view_hash dataset_id, table_id, table_name, description }
   let(:view_gapi) { Google::Apis::BigqueryV2::Table.from_json view_hash.to_json }
-  let(:view) { Google::Cloud::Bigquery::View.from_gapi view_gapi,
+  let(:view) { Google::Cloud::Bigquery::Table.from_gapi view_gapi,
                                                 bigquery.service }
